pyproject.toml
[project]
name = "pii-extractor"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "agno>=2.4.3",
    "fastapi>=0.128.0",
    "gliner>=0.2.24",
    "ipykernel>=7.1.0",
    "numpy>=2.4.1",
    "openai>=2.15.0",
    "pandas>=3.0.0",
    "protobuf>=6.33.4",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "torch>=2.10.0",
    "transformers>=4.57.6",
]
